Handle newly-created twig files without restarting
